[PowerMaster] v0.3

+++ Cluster/PowerMaster/src/powermaster/SolverThread.java
@@ -44,7 +44,7 @@ public class SolverThread extends Thread {
         return solver.getPopulation();
     }
     
-    public Collection<Individual> getUniqueIndividuals(int fitness){
+    public Collection<Individual> getUniqueIndividuals(double fitness){
         return PopulationUtils.getUniqueIndividuals(this.solver.getPopulation(),fitness);
     }
